fix imports in variational_algorithms
